+++ CHANGES.md
@@ -1,9 +1,14 @@
 # Changelog
 
+## version 1.2.0
+
+- change coreseVersion() to engineVersion()
+- add more unit tests
+
 ## Version 1.0.1
 
 - document local installaiton process (INSTALL.md)
-- build jar file then running `python -m build` (may need more work)
+- build jar file then running `python -m build` (may need more work)b
 - add doc skeleton and github action files
 - add coreseVersion (not fully implemented yet)